pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "antmeta"
version = "0.1.0"
description = "蚁元 - 安全的 OpenClaw 多代理团队系统 | Safe multi-agent team system for OpenClaw"
readme = {file = "README.md", content-type = "text/markdown"}
license = {file = "LICENSE"}
authors = [
    {name = "AntMeta Contributors"}
]
keywords = [
    "openclaw",
    "multi-agent",
    "agent-team",
    "ai-agents",
    "collaboration",
    "swarm-intelligence",
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/AntMeta/openclaw-antmeta"
Documentation = "https://github.com/AntMeta/openclaw-antmeta/blob/main/docs"
Repository = "https://github.com/AntMeta/openclaw-antmeta"
Issues = "https://github.com/AntMeta/openclaw-antmeta/issues"

[project.scripts]
antmeta = "antmeta.cli:main"

[tool.setuptools]
packages = ["antmeta"]
package-dir = {"" = "src"}

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
